Plotting random effects

This is the plotting method for random effects (simple random intercepts).

# S3 method for random.effect
plot(x, trans = identity, ...)

Arguments

x
a random effect object, extracted using mgcViz::sm.

trans
monotonic function to apply to the fit, confidence intervals and residuals, before plotting. Monotonicity is not checked.

...
currently unused.

Value

An object of class plotSmooth.

Examples

# NOT RUN {
library(mgcv)
b <- gam(travel~s(Rail,bs="re"), data=Rail, method="REML")
b <- getViz(b)
plot(sm(b, 1)) + l_fitLine(colour = 2, linetype = 2) + l_points() +
  l_ciLine(colour = 4, linetype = 3)

plot(sm(b, 1)) + l_ciPoly() + l_points()

# Default
plot(b)

# }