Extracting a smooth effect from a GAM model

This function can be used to extract a smooth or random effect from an object of class gamViz.

sm(o, select)

Arguments

o
an object of class gamViz, the output of a getViz call.

select
index of the selected smooth or random effect.

Value

An object representing a smooth effect.

See also

See getViz for examples.